demo files
Dru might have just broken keil by accidentally changing files in his library

+++ queued.c
@@ -9,14 +9,14 @@ bool cq_stuff(queue_t *q0);
 
 void init_queue(queue_t *ptr_to_queue, int max_items)
 {
-  uint8_t * new_buff = (uint8_t *const) malloc((max_items+1) *sizeof(uint8_t));  
+  uint16_t * new_buff = (uint16_t *const) malloc((max_items+1) *sizeof(uint16_t));  
   ptr_to_queue ->buffer = new_buff;
   ptr_to_queue ->head = 0;
   ptr_to_queue ->tail = 0;
   ptr_to_queue ->max_index = max_items;
 }
 
-bool pop(queue_t *q0, uint8_t *data)
+bool pop(queue_t *q0, uint16_t *data)
 {
     bool success;
     if(q0->head == q0->tail)  // if empty
@@ -33,9 +33,9 @@ bool pop(queue_t *q0, uint8_t *data)
     return success;
 }
 
-bool push(queue_t *q0, uint8_t data)
+bool push(queue_t *q0, uint16_t data)
 {
-    uint8_t new_tail = q0->tail + 1;
+    uint16_t new_tail = q0->tail + 1;
     if (new_tail > q0->max_index) {new_tail = 0; } // loop back around if going past max index
     if (new_tail == q0->head) {return false; } //queue full = error
     else{
